Spec for complete and due tasks, and task items
This adds methods and specs for determining if a task is complete,
due, and whether it has items. The `to_s` method for Task also
shows if a task has a due date.

The output of the rudimentary `td` script now numbers tasks by
their order. Eventually, this will be used when manipulating tasks
through the app.